Refactor override to decorator
